+++ sdk/common/utils.js
@@ -166,11 +166,27 @@ const isNode = () => {
   );
 };
 
+/**
+ * Checks if the current execution context is a Web Worker (DedicatedWorker).
+ *
+ * This function determines whether the code is running inside a Dedicated Web
+ * Worker by verifying that the global `self` object exists, has a constructor,
+ * and that the constructor's name is "DedicatedWorkerGlobalScope".
+ *
+ * @returns {boolean} True if running in a Dedicated Web Worker, otherwise false.
+ */
+const isWebWorker = () =>
+  typeof self === "object" &&
+  self.constructor &&
+  self.constructor.name === "DedicatedWorkerGlobalScope";
+
 const convertStringToBase64 = (string) => {
   if (isNode()) {
     return Buffer.from(string, "utf-8").toString("base64");
   } else if (isBrowser()) {
     return window.btoa(string);
+  } else if (isWebWorker()) {
+    return BufferPolyFill.from(string, "utf-8").toString("base64");
   } else {
     throw new FDKException("Base64 conversion error: Unsupported environment");
   }
@@ -217,4 +233,5 @@ module.exports = {
   NAV_TYPE,
   combineURLs,
   isAbsoluteURL,
+  isWebWorker,
 };